services/gateway/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
pydantic==2.8.2
python-dotenv==1.0.1
paho-mqtt==2.1.0
# local path build in docker-compose context (optional)
# nesshash-adapter @ file:///app/../../libs/nesshash-adapter
